modify/move var in right structure
