+++ src/rag_prep_tool/preprocessing/clean_text.py
@@ -1,8 +1,6 @@
 import re 
 from typing import List 
 
-from rag_prep_tool.vars import ART_OF_HAPPINESS_CHAPTERS_PAGE_NUMBERS
-from fast_antx.core import transfer
 
 def remove_space(text:str)->str:
     text = text.strip().replace("\n","").replace(" ","")
